refactor: extract ListPagination to components/list
